+++ frontend/src/pages/Dashboard/Dashboard.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 import Sidebar from '../../components/Sidebar/Sidebar';
-import DriverTable from './DriversTable';
-import RiderTable from './RidersTable';
+import Drivers from './Drivers';
+import Riders from './Riders';
 import Home from './Home';
 import Settings from './Settings';
 import './dashboard.css';
@@ -12,8 +12,8 @@ const Dashboard = () => (
     <Sidebar>
       <Switch>
         <Route path="/home" component={Home} />
-        <Route path="/drivers" component={DriverTable} />
-        <Route path="/riders" component={RiderTable} />
+        <Route path="/drivers" component={Drivers} />
+        <Route path="/riders" component={Riders} />
         <Route path="/settings" component={Settings} />
         <Route path="*">
           <Redirect to="/home" />
