+++ frontend/src/pages/Dashboard/Drivers.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Table from '../../components/UserTables/DriversTable';
+
+const Drivers = () => (
+  <Table />
+);
+
+export default Drivers;
