+++ frontend/src/pages/Dashboard/Riders.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import Table from '../../components/UserTables/RidersTable';
+
+const Riders = () => (
+  <Table />
+);
+
+export default Riders;
