+++ frontend/src/types/index.ts
@@ -1,0 +1,25 @@
+export type AccessibilityNeeds = {
+  hasCrutches: boolean;
+  needsAssistant: boolean;
+  needsWheelchair: boolean;
+}
+
+export type Rider = {
+  id: string;
+  firstName: string;
+  lastName: string;
+  phoneNumber: string;
+  email: string;
+  accessibilityNeeds: AccessibilityNeeds;
+  description: string;
+  joinDate: string;
+  pronouns: string;
+  address: string;
+}
+
+export type Driver = {
+  name: string;
+  netid: string;
+  email: string;
+  phone: string;
+}
